API: fix C app build errors
Untangle some Rom-era weirdness:
- move boinc_resolve_filename(), boinc_resolve_filename_s()
    and resolve_soft_link()
    from lib/app_ipc.cpp to api/boinc_api.cpp
- remove the #ifdef __cplusplus from app_ipc.h
- add #ifdef __cplusplus in common_defs.h because we only
    need a little of it, and the rest is not C
